feature/MBX-3248-ExampleApp: add ChooseInappMessageDelegate
